% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesqele_read.R
\name{available_years}
\alias{available_years}
\title{List available years}
\usage{
available_years()
}
\value{
character vector of available years.
}
\description{
Obtain available years directly from TSE site.
}
