% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities}
\alias{cities}
\title{Cities list}
\format{A data frame with 5571 rows and 2 variables:
\describe{
  \item{uf}{State initials.}
  \item{muni}{City code.}
}}
\source{
\url{http://inter01.tse.jus.br/pesqele-publico/app/pesquisa/listarEstatisticos.xhtml}
}
\usage{
cities
}
\description{
A dataset containing a list of states and city codes from TSE pesqEle platform.
}
\keyword{datasets}

