% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conre3.R
\name{get_conre3_companies}
\alias{get_conre3_companies}
\title{Download CONRE-3 registered companies}
\usage{
get_conre3_companies()
}
\value{
\code{tibble} with 5 columns as specified in
  \code{\link{conre3_companies}}.
}
\description{
Downloads CONRE-3 registered companies from CONRE-3 sites.
}

