% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{pe_2018}
\alias{pe_2018}
\title{Downloads and parses all data from 2018 Elections}
\usage{
pe_2018(path = "data-raw/html_2018")
}
\arguments{
\item{path}{directory path to save files. Defaults to 
  \code{data-raw/html_2018}.

This function downloads all polls for 2018 elections. It does not delete 
existing polls, then it might run faster for the next times. It does
subscribe the \code{pesq_main} files, whose contain the list of individual
polls. This way you can always run this function pointing to the same 
directory, and it will update the polls safely, without downloading 
uninportant files. The full download takes about 15 minutes depending on
the internet connection.}
}
\value{
\code{tibble} contaning the same variables as the  
  \code{\link{pesqEle_2018}} database.
}
\description{
Downloads HTML files from
\href{http://inter01.tse.jus.br/pesqele-publico/}{Brazilian Superior Electoral Court platform}.
Then parses all the HTML files, returning a tidy dataset.
}
