% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar.R
\name{pesq_city_codes}
\alias{pesq_city_codes}
\title{List of cities}
\usage{
pesq_city_codes()
}
\value{
\code{tibble} with two columns, state name and city code.
There's a cached result in \code{\link{cities}}.
}
\description{
Downloads list of series according to TSE.
It is obtained automatically via web scraping.
Needs stable internet connection.
}
\seealso{
\code{\link{cities}}.
}

