% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pesq_details}
\alias{pesq_details}
\title{pesqEle details results}
\format{A data frame with 210347 rows and 3 variables:
\describe{
   \item{arq}{path to the HTML file.}
   \item{key}{Name of the info.}
   \item{val}{Value of the info.}
}}
\source{
\url{http://inter01.tse.jus.br/pesqele-publico/app/pesquisa/listarEstatisticos.xhtml}
}
\usage{
pesq_details
}
\description{
A raw dataset containing the details of each result.
}
\keyword{datasets}
