% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pesq_main}
\alias{pesq_main}
\title{pesqEle main results}
\format{A data frame with 11078 rows and 8 variables:
\describe{
   \item{arq}{path to the HTML file.}
   \item{numero_de_identificacao}{Research id.}
   \item{empresa_contratada_nome_fantasia}{Fantasy name of the company.}
   \item{no_do_estatistico}{Statistician id.}
   \item{nome_do_estatistico}{Statistician name.}
   \item{data_de_registro}{Registry date.}
   \item{abrangencia}{State / city.}
   \item{acoes}{HTML buttons.}
}}
\source{
\url{http://inter01.tse.jus.br/pesqele-publico/app/pesquisa/listarEstatisticos.xhtml}
}
\usage{
pesq_main
}
\description{
A raw dataset containing the results of all the city searches.
Cities with zero results are informed but have the column
\code{numero_de_identificacao} equal 'Nenhum registro encontrado!'.
}
\keyword{datasets}

