% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{pesq_parse_details}
\alias{pesq_parse_details}
\title{Parse details}
\usage{
pesq_parse_details(arqs)
}
\arguments{
\item{arqs}{character vector containing file paths. Valid files have \code{<UF>_<code>_<result_id>.html} format.}
}
\value{
\code{tibble} containing 3 columns.
}
\description{
Parse details HTML files returned by \code{\link{pesq_download_cities}}
into \code{tibbles}.
}
