% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{pesq_parse_main}
\alias{pesq_parse_main}
\title{Parse main results}
\usage{
pesq_parse_main(arqs)
}
\arguments{
\item{arqs}{character vector containing file paths. Valid files have \code{<UF>_<code>.html} format.}
}
\value{
\code{tibble} containing 8 columns.
}
\description{
Parse main HTML files returned by \code{\link{pesq_download_cities}}
into \code{tibbles}.
}
