% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pesqEle}
\alias{pesqEle}
\title{pesqEle tidy dataset}
\format{A data frame with 11078 rows and 8 variables:
\describe{
   \item{id_seq}{.}
   \item{id_pesq}{.}
   \item{id_muni}{.}
   \item{info_uf}{.}
   \item{info_muni}{.}
   \item{info_election}{.}
   \item{info_position}{.}
   \item{comp_nm}{.}
   \item{comp_cnpj}{.}
   \item{comp_contract_same}{.}
   \item{stat_id}{.}
   \item{stat_nm}{.}
   \item{pesq_n}{.}
   \item{pesq_val}{.}
   \item{pesq_origin}{.}
   \item{pesq_contractors}{.}
   \item{dt_reg}{.}
   \item{dt_pub}{.}
   \item{dt_start}{.}
   \item{dt_end}{.}
   \item{txt_verif}{.}
   \item{txt_method}{.}
   \item{txt_about}{.}
   \item{txt_plan}{.}
}}
\source{
\url{http://inter01.tse.jus.br/pesqele-publico/app/pesquisa/listarEstatisticos.xhtml}
}
\usage{
pesqEle
}
\description{
A tidy dataset containing the results of elections polls.
}
\keyword{datasets}
