% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesqele_read.R
\name{pesqele_fetch}
\alias{pesqele_fetch}
\title{Download and read data from TSE}
\usage{
pesqele_fetch(year = available_years(), uf = "all")
}
\arguments{
\item{year}{character vector of election years.
Defaults to \code{\link[=available_years]{available_years()}}.}

\item{uf}{character vector of states in UF code, like \code{c("AC", "BA")}.
Defaults to \code{all}, which gets all available UFs.}
}
\description{
Download and read data from TSE
}
\seealso{
\code{\link[=available_years]{available_years()}}
}
