% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesqele_tidy.R
\name{pesqele_tidy}
\alias{pesqele_tidy}
\title{Clean pesqEle data}
\usage{
pesqele_tidy(da_pesqele)
}
\arguments{
\item{da_pesqele}{Data obtained using \code{\link[=pesqele_fetch]{pesqele_fetch()}}.}
}
\description{
Clean pesqEle data using simple heuristics. These heuristics may be
incomplete and sometimes even wrong. To have complete control of what
you're doing, consider using the raw data.
}
