% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ufs}
\alias{ufs}
\title{Brazil state codes}
\format{A data frame with 27 rows and 2 variables:
\describe{
   \item{uf}{State initials}
   \item{CD_GEOCUF}{State code}
}}
\source{
IBGE.
}
\usage{
ufs()
}
\description{
A tidy dataset containing Brazil state codes.
}
